different test files for eth and aion tests
