+++ packages/web3-provider/package.json
@@ -8,7 +8,7 @@
         "tests": "./tests"
     },
     "scripts": {
-        "test": "mocha tests/providerTest.js"
+        "test": "mocha tests/*.js"
     },
     "repository": "https://github.com/pokt-network/pocket-js/tree/master/packages/web3_provider",
     "homepage": "https://github.com/pokt-network/pocket-js",
